Regression Models
K nearest neighbors, Random forest and multiple linear regression have been made for the regression models. I tried setting the weights parameter to distance which led to a better test score than having the weight parameter to uniform however this lead to over fitting of the training data and an accuracy of 100. Most of the accuracies are around 80% but I would like to have them better. I'll try maybe correcting the data and seeing if that chances anything. I can also look into adding weights to the input parameters.